symbolizer: refactor to build on different platform
Before this patch, the symbolizer package only builds on Linux. In
preparation for making the ustack operator work client-side, this patch
moves part of the code in _linux.go files. It will make it easier for
future patches.
